# Add NonderterministicFastCommit to speed up migrations when ordering isn't required