@fxamacker 2024-05-10
Looks like validation and storage health checks passed tonight with this PR added to atree migration program at:
- (locally modified) https://github.com/onflow/flow-go/tree/feature/atree-inlining-cadence-v0.42

Need to take another look at logs to confirm.