@fxamacker 2024-05-13
> Would it be possible to get a diff of this new function and the existing deterministic version of it? It would make it easier to review, by seeing how they differ.
>
> If the functions only differ in a small way, maybe it's possible to extract the shared code and only make the parts that differ configurable?

@turbolent  Good suggestions!  However, in this case there are more changes than unchanged code so a diff, etc. isn't as helpful as it normally would be.

The changes are mostly about eliminating overhead and parallelizing as much as possible.  

I outlined steps in both functions below so we can see the differences at higher level.

`FastCommit()` commits deltas in deterministic order:
- sort non-temp slab IDs by address and index
- send non-temp slab IDs to job queue
- (in goroutines) read and encode slabs from deltas
  - job queue (input): slab ID
  - result queue (output): encoded slab and possible error
- collect all encoded results from result queue in temporary map
- commit deleted and encoded results in order to base storage

In order to commit in deterministic order, the commit step has to wait until all slabs are encoded, and there are also overheads from sorting slab IDs and collecting encoded results in temporary map.

`NonderterministicFastCommit()` commits deltas in nondeterministic order:
- send non-temp and non-empty slab to job queue
- (in goroutines) read and encode slabs 
  - job queue (input): slab ID and slab
  - result queue (output): encoded slab and possible error
- commit deleted slabs to base storage  
- commit encoded slabs from result queue to base storage

In nondeterministic approach, the commit steps are in parallel with encoding and there isn't any temp map to collect data.

Even though goroutine encoding step is similar, `FastCommit()` reads slab from `deltas` map while `NonderterministicFastCommit()` reads slab from job queue.  Although I can modify `FastCommit()` to be the same in this aspect, it doesn't seem worthwhile since that would use more memory without making `FastCommit()` faster.